Rearranging and using headers files like I'm supposed to.
Since I decided to restart this doing it the correct way I added header files.
So far tsp.c compiles and works correctly. Now I just need to add in the list
stuff to handle to creation of tours and the stack. From there I should be able
to use this as a basis for doing the recursive DFS, iterative DFS, and the
second iteration presented in the book which I didn't look at much.